updating adv task 3 no 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,17 +11,17 @@
 
 int main(void)
 {
-	int a = 1;
-	int b = 2;
-	int c;
+	long int a = 1;
+	long int b = 2;
+	long int c;
 	int d;
 
-	printf("%d, %d", a, b);
+	printf("%lu, %lu", a, b);
 
 	for (d = 2; d < 50; d++)
 	{
 		c = a + b;
-		printf(", %d", c);
+		printf(", %lu", c);
 		a = b;
 		b = c;
 	}
